ModelTraining/examples/example_model/requirements.txt
adal==1.2.4
applicationinsights==0.11.9
azure-common==1.1.25
azure-core==1.8.2
azure-graphrbac==0.61.1
azure-identity==1.4.0
azure-mgmt-authorization==0.61.0
azure-mgmt-containerregistry==2.8.0
azure-mgmt-keyvault==2.2.0
azure-mgmt-resource==10.2.0
azure-mgmt-storage==11.2.0
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
azureml-automl-core==1.16.0
azureml-contrib-dataset==1.15.0
azureml-core==1.16.0.post1
azureml-dataprep==2.3.4
azureml-dataprep-native==23.0.0
azureml-dataprep-rslex==1.1.3
azureml-dataset-runtime==1.16.0
azureml-pipeline==1.16.0
azureml-pipeline-core==1.16.0
azureml-pipeline-steps==1.16.0
azureml-sdk==1.16.0
azureml-telemetry==1.16.0
azureml-train==1.16.0
azureml-train-automl-client==1.16.0
azureml-train-core==1.16.0
azureml-train-restclients-hyperdrive==1.16.0
backcall==0.2.0
backports.tempfile==1.0
backports.weakref==1.0.post1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cloudpickle==1.6.0
contextlib2==0.6.0.post1
cryptography==3.1.1
decorator==4.4.2
distro==1.5.0
docker==4.3.1
dotnetcore2==2.1.16
fusepy==3.0.1
idna==2.10
importlib-metadata==2.0.0
ipykernel==5.3.4
ipython==7.16.1
ipython-genutils==0.2.0
isodate==0.6.0
jedi==0.17.2
jeepney==0.4.3
jmespath==0.10.0
joblib==0.17.0
jsonpickle==1.4.1
jupyter-client==6.1.7
jupyter-core==4.6.3
msal==1.5.0
msal-extensions==0.2.2
msrest==0.6.19
msrestazure==0.6.4
ndg-httpsclient==0.5.1
numpy==1.19.2
oauthlib==3.1.0
pandas==1.1.3
parso==0.7.1
pathspec==0.8.0
pexpect==4.8.0
pickleshare==0.7.5
portalocker==1.7.1
prompt-toolkit==3.0.8
ptyprocess==0.6.0
pyarrow==1.0.1
pyasn1==0.4.8
pycparser==2.20
Pygments==2.7.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
python-dateutil==2.8.1
pytz==2020.1
pyzmq==19.0.2
requests==2.24.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
scikit-learn==0.23.2
scipy==1.5.2
SecretStorage==3.1.2
six==1.15.0
sklearn==0.0
threadpoolctl==2.1.0
tornado==6.0.4
traitlets==4.3.3
urllib3==1.25.10
wcwidth==0.2.5
websocket-client==0.57.0
zipp==3.3.0
